projectFiles/HARMONIE forecast/parseSaveGribHarmonie.gcloud/requirements.txt
# Function dependencies, for example:
# package>=version
pupygrib
requests
pytz
google-cloud-storage
python-dateutil
lxml
cssselect